1. entity설정 할 때 name에 db 컬럼명과 맞춰줘야함 안 맞출 시 fatG 같은 카멜케이스는 자동으로 스네이크 변경에 실패해서 fatg로 저장되어 오류 남 2. 모든 엔티티의 고유 id의 타입을 int로 바꿔서 db sql과 대응되게 수정

+++ backend/src/main/java/com/yamyam/diet/entity/CourseSchedule.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class CourseSchedule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long scheduleId;
+    private int scheduleId;
 
     @Column(nullable = false)
     private LocalDate scheduleDate;
@@ -24,7 +24,7 @@ public class CourseSchedule {
     private List<ScheduleDish> scheduleDishes;
 
     // Getters
-    public Long getScheduleId() {
+    public int getScheduleId() {
         return scheduleId;
     }
 
@@ -45,7 +45,7 @@ public class CourseSchedule {
     }
 
     // Setters
-    public void setScheduleId(Long scheduleId) {
+    public void setScheduleId(int scheduleId) {
         this.scheduleId = scheduleId;
     }
